Refactoring rules.  Using a factory now to produce the rules object.
